[coordinator] Move coordinator_control to standalone directory.
